extracts props types
