+++ BookCatalog.API/Controllers/BookController.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 namespace BookCatalog.API.Controllers
 {
     [Route("api/[controller]")]
-    //[Authorize]
+    [Authorize]
     [ApiController]
     public class BookController : ControllerBase
     {
